fixed trpc on server side
